stt-service/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
numpy
python-multipart
